requirements.txt
asgiref==3.3.1
Django==3.1.7
django-debug-toolbar==3.2
django-debug-toolbar-force==0.1.8
django-nine==0.2.4
django-parler==2.2
django-parler-rest==2.1
djangorestframework==3.12.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
